Add makefile

+++ block.cpp
@@ -9,7 +9,7 @@ char Block::typeToSign(Block::BlockType t)
 
 Block::BlockType Block::signToType(char c)
 {
-    for ( int i = 0 ; i < strlen(signs) ; i++ )
+    for ( size_t i = 0 ; i < strlen(signs) ; i++ )
         if ( signs[i] == c )
             return BlockType(i);
     throw UnknownType();
